requirements.txt
tensorflow>=2.11
numpy
pandas
scikit-learn
matplotlib
streamlit
pillow
opencv-python
tqdm
